Update epics index JSON response
For JSON requests epics index page returns a JSON array of epics. This
will be used on epic roadmap page (#3559). Pagination is disabled in
this case - it's not supported for the first iteration on roadmap page.

Until now JSON response for epics index action returned an HTML list of
epics but it was not used anywhere in our code.

Related #3559