Make GITLAB_FEATURES in build_spec compatible with EE
Right now this fails in EE since the EE codebase defaults to having a starter license which means there are features in this value. This seems to be incompatible with the statement that tests should behave the same in CE as EE but I think that since there is a before on all EE specs creating a starter license it unfortunately does not behave that way.
This is needed for #40994 and fixes the failing specs in https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/4393