make sure there is a dependency on Gitlab::CurrentSettings is
This fixes an issue where the Rails autoload system would throw various `Unable to autoload constant` errors (such as `Unable to autoload constant EE::ProjectsHelper`) when using the autoload system (such with `spring` or `reload!` in the rails console.

This error was specifically ocurring in the EE code, however, it's seems reasonable to place the fix in CE as a general innoculation.