Add partial index on projects for index-only scans.
This helps with queries that get project ids based on the - comparably
rare - visibility levels 10 and 20. For these, postgres can now leverage
the partial index for a index-only scan to improve performance.

Example queries:
SELECT id FROM projects WHERE visibility_level IN (10,20)
SELECT id FROM projects WHERE visibility_level IN (10)

For MySQL, this results in a full index on id because MySQL omits the
WHERE clause. That is, the index is a duplicate of the primary key
basically.