Workaround for epic spec failure
For some reason when a certain set of specs are run together, the input to the
epic field gets stuck waiting for more text. Clicking the "Add" button only
causes it to get out of this state, but it doesn't actually submit the
request. By sending a keystroke before clicking the button, we appear to get
more consistent behavior.

Closes #4185