Fix spec/features/projects/clusters/applications_spec.rb (#42643)>
Also remove the flakiness by using
`expect(page).to have_css(..., exact_text: '')` instead of
`expect(page.find(...).text).to eq('')` since the former will wait until
the text has this value while the latter will only wait for the element
to exist then test its text straight away which introduces a race
condition if the element is already there with a previous text value.
This test was failing locally most of the time because of this other
way of asserting.