CONTRIBUTING.md: Use a proper example for stewardship label
The bug guidelines example of what should be marked with the stewardship
label goes against the principles of GitLab as stated in
https://about.gitlab.com/stewardship/ since mentions removing a feature
from GitLab CE to make it exclusive of GitLab EE.

The example should be the opposite, where GitLab Inc. is bringing a
feature from GitLab EE to GitLab CE.

This commit changes the example to reflect this situation.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/43579