Bug.md: Use html comments instead of asking to remove the notice
The issue template for bugs has a notice at the start that mentions to
search for duplicates before creating a new issue. Then the reporter
is asked to remove the notice once read, presumptively to avoid noise
in the actual issue.

Instead of that, we can make that part of the issue template an html
comment, so it won't be rendered in the final result. This avoids
the tedious task of removing that part of the notice.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/43580