Remove use of any_instance_of in runner spec
In EE we prepend a module into BuildFinishedWorker. This would cause
this spec to fail because RSpec doesn't support the use of
"any_instance_of" with classes that include a prepended module.

To work around this we just stub the "perform_async" class method
instead of stubbing the "perform" instance method.