Apply query changes to snippets only.
This applies the changes introduced in `Project.public_or_visible_to_user`
to the snippets finder only.

We know that for `SnippetsFinder`, this change improves SQL timing from
5/23/25s to 0.7/2/4s (mean/p95/p99). At the same time, the scope was too
broad, (negatively) affecting SQL timings in various other places:

https://gitlab.com/gitlab-com/infrastructure/issues/3772

With this commit, the snippets dashboard stays usuable as we generally
don't run into statement timeouts. In contrast to the earlier change in
!17088, the scope of the change is limited to `SnippetsFinder` only,
thus not affecting other places.