Optional '/-/' delimiter for epics and search API
'/-/' delimiter is used only in UI, in API we don't use it for
other endpoints. To align epics and search endpoints with the rest of API
endpoints, this patch makes '/-/' optional for existing endpoints (to
keep backward compatibility), for new endpoints (epic notes, discussions)
'/-/' path is not included.

Documentation is updated to prefer paths without '/-/'.

Closes #5245