Fix audit and Geo project deletion events not being logged under certain conditions
This commit fixes two issues:

1. Due to DNS errors, one of the after_commit hooks were causing an
Excon::SocketError error, which then got caught by the exception handler in
Projects::DestroyService. However, since the model was already deleted in the
database, attempting to update it would fail with a "Can't modify frozen hash"
error.

2. The Projects::DestroyService would then return `false`, which would cause
the prepended EE::Projects::DestroyService to skip logging of the project
deletion. Now we log an event if we see that the project was destroyed.

Closes gitlab-org/gitlab-ce#44378