Make `ci_cd_projects` globally available for .com
The coming year `ci_cd_projects` should be available for _all_
projects on GitLab.com. So making them global features allows all
projects to use them.

For that to work, we also need to enable `repository_mirrors` and
`github_project_service_integration` so the projects can correctly be
mirrored to allow CI to run.