Force Rails to not complain about reloading
Same strategy with:
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/17810

See:
https://stackoverflow.com/a/29710188/1992201

Frankly I don't really understand how this works and I don't really
care either. However I tried it and it does the job. To try this,
make sure you have pending migrations, and run the server, hit the site.
It would tell you that there's pending migrations, and then run
migrations, and then hit the site again.

Without this patch, Rails would complain that "A copy of ...",
with this patch, it works without problems.