Fix enviroments seeding on certain branch names
Bit of a cryptic commit title, but our seeding will try to guard against
branchnames being anything but alphanumerical. These charachters we gsub
away but can't have an empty string as replacement.

This change is less efficient, but given that the number of seeded MRs
per project is very limited, this should be ok.