edit GitLab license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit updates the LICENSE file so that it contains the exact text
of the MIT license. It also moves the reference to third-party software
licensing to the README.

These changes allow Licensee to successfully identify the license
type of GitLab's codebase as MIT.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>