Check Gitaly when running `rake dev:setup`
Before this change, trying to setup dev environment could be tried
without having at least one Gitaly running. Cloning the repositories
would fail, but not stop the setup.

Given this would lead to an inconsistent state, a check was added if we
could connect to the server.

Output when it fails:

```
$ rake dev:setup
Failed to connect to Gitaly...
Error: 14:Connect Failed
```