Port of jr-33320-lfs-settings-interface to EE
LFS can be enabled and disabled per project using the API by Masters
and Owners, but the UI is only available to Admins. The interface and
API should be consistent.

LFS can also be enabled and disabled per group using the API by Owners,
but the UI is only available to Admins. This interface should also be
consistent.

Additionally removes an unneeded check if 2FA interface, since the
interface is only available to Owners.

Closes #33320