Fix DB LB errors when escaping input
The Host class uses the pattern `connection.quote(input)` to quote
certain forms of input before passing these to queries. Rails will
unfortunately try to acquire a database connection when running such
code, resulting in errors should the database not be available.

This commit fixes this by simply catching the connection errors that may
be produced.